pyproject.toml
[tool.poetry]
name = "labeler"
version = "1.0.0"
description = ""
authors = ["Binh Vu <binh@toan2.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hugedict = "^2.12.9"
streamlit = "^1.31.1"
pandas = "^2.2.0"
serde2 = "^1.7.3"
python-slugify = "^8.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
